repo-step-3-01/pyproject.toml
[project]
name = "repo-step-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.72.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]
